core/requirements.txt
argcomplete==2.0.0
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
boto3==1.26.26
botocore==1.29.26
certifi==2022.12.7
cfn-flip==1.3.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
djangorestframework==3.14.0
drf-yasg==1.21.4
durationpy==0.5
exceptiongroup==1.0.4
flake8==6.0.0
hjson==3.1.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==22.0
pathspec==0.10.2
phonenumbers==8.13.2
placebo==0.9.0
platformdirs==2.6.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.0
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
troposphere==4.2.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
zappa==0.56.1
